requirements.txt
json-stream==2.3.3
requests==2.31.0
python-dotenv==1.0.1
boto3>=1.26.0
